feature 2410 added: uploadify upgraded from version 2.1.0 to 3.0.0 (still beta, but said as stable by users on uploadify forums)
bug 2411 fixed: thanks to uploadify 3.0.0, the "Browse" button is now localized
(text in any language)

bug 1653 fixed: if the selected file exceeds the upload_max_size (as defined in
the php.ini), an alert is shown and the file is not added to the queue.

bug 2412 fixed: ability to select GIF files with the Flash Uploader.