merge r26952 from branch 2.6 to trunk
bug 3035 fixed: the output of pwg.categories.getList with option
tree_output=true was modified in r22729 to fix a bug. But fixing this bug has
broken iOS/Android applications which uses this parameter. The immediate
solution is to reintroduce the bug for now and fix it cleanly later (requires
a new version of iOS/Android apps)